+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+-   Reworded SCS103 and extend it to include a few more cases:
+    + `subprocess.getoutput()`
+    + `subprocess.getstatusoutput()`
+    + `asyncio.create_subprocess_shell()`
+    + `loop.subprocess_shell()`
+
 ## [1.2.0] - 2021-07-19
 
 -   Added SCS110 to avoid using `os.popen()` as it internally uses `subprocess.Popen` with `shell=True`